# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorMessage {
  error: string;
}

export type PoseCategory = typeof PoseCategory[keyof typeof PoseCategory];


export const PoseCategory = {
  Centering: 'Centering',
  'Warm-Up': 'Warm-Up',
  Standing: 'Standing',
  Backbend: 'Backbend',
  Twist: 'Twist',
  Hip: 'Hip',
  Floor: 'Floor',
  Rest: 'Rest',
  Closing: 'Closing',
} as const;

export type PoseDurationType = typeof PoseDurationType[keyof typeof PoseDurationType];


export const PoseDurationType = {
  time: 'time',
  breaths: 'breaths',
} as const;

export type PoseCautionsItem = typeof PoseCautionsItem[keyof typeof PoseCautionsItem];


export const PoseCautionsItem = {
  Back: 'Back',
  Knees: 'Knees',
  Wrists: 'Wrists',
  Neck: 'Neck',
  Shoulders: 'Shoulders',
  Hips: 'Hips',
  Balance: 'Balance',
} as const;

export interface Pose {
  id: number;
  name: string;
  category: PoseCategory;
  defaultDurationSeconds: number;
  durationType: PoseDurationType;
  /** @nullable */
  defaultBreaths: number | null;
  perSide: boolean;
  cue: string;
  cautions: PoseCautionsItem[];
  modification: string;
  chairOption: string;
  isCustom: boolean;
}

export type PoseInputCategory = typeof PoseInputCategory[keyof typeof PoseInputCategory];


export const PoseInputCategory = {
  Centering: 'Centering',
  'Warm-Up': 'Warm-Up',
  Standing: 'Standing',
  Backbend: 'Backbend',
  Twist: 'Twist',
  Hip: 'Hip',
  Floor: 'Floor',
  Rest: 'Rest',
  Closing: 'Closing',
} as const;

export type PoseInputDurationType = typeof PoseInputDurationType[keyof typeof PoseInputDurationType];


export const PoseInputDurationType = {
  time: 'time',
  breaths: 'breaths',
} as const;

export type PoseInputCautionsItem = typeof PoseInputCautionsItem[keyof typeof PoseInputCautionsItem];


export const PoseInputCautionsItem = {
  Back: 'Back',
  Knees: 'Knees',
  Wrists: 'Wrists',
  Neck: 'Neck',
  Shoulders: 'Shoulders',
  Hips: 'Hips',
  Balance: 'Balance',
} as const;

export interface PoseInput {
  /** @minLength 1 */
  name: string;
  category: PoseInputCategory;
  /** @minimum 5 */
  defaultDurationSeconds: number;
  durationType: PoseInputDurationType;
  /** @nullable */
  defaultBreaths?: number | null;
  perSide: boolean;
  cue: string;
  cautions: PoseInputCautionsItem[];
  modification: string;
  chairOption: string;
}

export interface TagLabel {
  id: number;
  name: string;
  isCustom: boolean;
}

export interface TagLabelInput {
  /** @minLength 1 */
  name: string;
}

export interface RoutineEntry {
  poseId: number;
  durationSeconds: number;
  /** @nullable */
  breaths: number | null;
}

export interface RoutineSections {
  centering: RoutineEntry[];
  flow: RoutineEntry[];
  closing: RoutineEntry[];
}

export interface Routine {
  id: number;
  title: string;
  description: string;
  tags: string[];
  sections: RoutineSections;
  totalSeconds: number;
}

export interface RoutineInput {
  /** @minLength 1 */
  title: string;
  description?: string;
  tags?: string[];
  sections: RoutineSections;
}

export interface RoutineSummary {
  id: number;
  title: string;
  description: string;
  tags: string[];
  totalSeconds: number;
  poseCount: number;
  cautions: string[];
}