# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: poses
    description: Pose bank operations
  - name: routines
    description: Routine operations
  - name: tags
    description: Tag operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /poses:
    get:
      operationId: listPoses
      tags: [poses]
      summary: List all poses in the pose bank
      responses:
        "200":
          description: List of poses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Pose"
    post:
      operationId: createPose
      tags: [poses]
      summary: Add a custom pose to the pose bank
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PoseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Pose"
  /tag-labels:
    get:
      operationId: listTags
      tags: [tags]
      summary: List all routine tags
      responses:
        "200":
          description: List of tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TagLabel"
    post:
      operationId: createTag
      tags: [tags]
      summary: Add a custom tag
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagLabelInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagLabel"
  /routines:
    get:
      operationId: listRoutines
      tags: [routines]
      summary: List all routines with summary info
      responses:
        "200":
          description: List of routine summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoutineSummary"
    post:
      operationId: createRoutine
      tags: [routines]
      summary: Create a routine
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoutineInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Routine"
  /routines/{id}:
    get:
      operationId: getRoutine
      tags: [routines]
      summary: Get a routine with its full sections
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Routine detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Routine"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
    put:
      operationId: updateRoutine
      tags: [routines]
      summary: Update a routine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoutineInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Routine"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
    delete:
      operationId: deleteRoutine
      tags: [routines]
      summary: Delete a routine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
  /routines/{id}/duplicate:
    post:
      operationId: duplicateRoutine
      tags: [routines]
      summary: Duplicate a routine
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Duplicated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Routine"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorMessage:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Pose:
      type: object
      required:
        [
          id,
          name,
          category,
          defaultDurationSeconds,
          durationType,
          defaultBreaths,
          perSide,
          cue,
          cautions,
          modification,
          chairOption,
          isCustom,
        ]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
          enum:
            [
              Centering,
              Warm-Up,
              Standing,
              Backbend,
              Twist,
              Hip,
              Floor,
              Rest,
              Closing,
            ]
        defaultDurationSeconds:
          type: integer
        durationType:
          type: string
          enum: [time, breaths]
        defaultBreaths:
          type: ["integer", "null"]
        perSide:
          type: boolean
        cue:
          type: string
        cautions:
          type: array
          items:
            type: string
            enum: [Back, Knees, Wrists, Neck, Shoulders, Hips, Balance]
        modification:
          type: string
        chairOption:
          type: string
        isCustom:
          type: boolean
    PoseInput:
      type: object
      required:
        [
          name,
          category,
          defaultDurationSeconds,
          durationType,
          perSide,
          cue,
          cautions,
          modification,
          chairOption,
        ]
      properties:
        name:
          type: string
          minLength: 1
        category:
          type: string
          enum:
            [
              Centering,
              Warm-Up,
              Standing,
              Backbend,
              Twist,
              Hip,
              Floor,
              Rest,
              Closing,
            ]
        defaultDurationSeconds:
          type: integer
          minimum: 5
        durationType:
          type: string
          enum: [time, breaths]
        defaultBreaths:
          type: ["integer", "null"]
        perSide:
          type: boolean
        cue:
          type: string
        cautions:
          type: array
          items:
            type: string
            enum: [Back, Knees, Wrists, Neck, Shoulders, Hips, Balance]
        modification:
          type: string
        chairOption:
          type: string
    TagLabel:
      type: object
      required: [id, name, isCustom]
      properties:
        id:
          type: integer
        name:
          type: string
        isCustom:
          type: boolean
    TagLabelInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
    RoutineEntry:
      type: object
      required: [poseId, durationSeconds, breaths]
      properties:
        poseId:
          type: integer
        durationSeconds:
          type: integer
        breaths:
          type: ["integer", "null"]
    RoutineSections:
      type: object
      required: [centering, flow, closing]
      properties:
        centering:
          type: array
          items:
            $ref: "#/components/schemas/RoutineEntry"
        flow:
          type: array
          items:
            $ref: "#/components/schemas/RoutineEntry"
        closing:
          type: array
          items:
            $ref: "#/components/schemas/RoutineEntry"
    Routine:
      type: object
      required: [id, title, description, tags, sections, totalSeconds]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        sections:
          $ref: "#/components/schemas/RoutineSections"
        totalSeconds:
          type: integer
    RoutineInput:
      type: object
      required: [title, sections]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        sections:
          $ref: "#/components/schemas/RoutineSections"
    RoutineSummary:
      type: object
      required: [id, title, description, tags, totalSeconds, poseCount, cautions]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
        totalSeconds:
          type: integer
        poseCount:
          type: integer
        cautions:
          type: array
          items:
            type: string