# lib/api-zod/src/generated/types/pose.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PoseCategory } from './poseCategory';
import type { PoseCautionsItem } from './poseCautionsItem';
import type { PoseDurationType } from './poseDurationType';

export interface Pose {
  id: number;
  name: string;
  category: PoseCategory;
  defaultDurationSeconds: number;
  durationType: PoseDurationType;
  /** @nullable */
  defaultBreaths: number | null;
  perSide: boolean;
  cue: string;
  cautions: PoseCautionsItem[];
  modification: string;
  chairOption: string;
  isCustom: boolean;
}