# lib/api-zod/src/generated/types/poseInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PoseInputCategory } from './poseInputCategory';
import type { PoseInputCautionsItem } from './poseInputCautionsItem';
import type { PoseInputDurationType } from './poseInputDurationType';

export interface PoseInput {
  /** @minLength 1 */
  name: string;
  category: PoseInputCategory;
  /** @minimum 5 */
  defaultDurationSeconds: number;
  durationType: PoseInputDurationType;
  /** @nullable */
  defaultBreaths?: number | null;
  perSide: boolean;
  cue: string;
  cautions: PoseInputCautionsItem[];
  modification: string;
  chairOption: string;
}