# lib/api-zod/src/generated/types/routine.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RoutineSections } from './routineSections';

export interface Routine {
  id: number;
  title: string;
  description: string;
  tags: string[];
  sections: RoutineSections;
  totalSeconds: number;
}