# lib/api-zod/src/generated/types/routineSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RoutineSummary {
  id: number;
  title: string;
  description: string;
  tags: string[];
  totalSeconds: number;
  poseCount: number;
  cautions: string[];
}