# lib/api-zod/src/generated/types/tagLabel.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TagLabel {
  id: number;
  name: string;
  isCustom: boolean;
}