# lib/api-zod/src/generated/types/tagLabelInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TagLabelInput {
  /** @minLength 1 */
  name: string;
}